ai-video-platform/requirements.txt
# Core Dependencies
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
httpx==0.26.0

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pymongo==4.6.1
redis==5.0.1
alembic==1.13.1

# AI/ML Dependencies
torch==2.2.0
torchvision==0.17.0
transformers==4.37.2
opencv-python==4.9.0.80
numpy==1.26.4
scipy==1.12.0
librosa==0.10.1
moviepy==1.0.3
ffmpeg-python==0.2.0

# Video Processing
ffmpeg-python==0.2.0
av==11.0.0
pytube==15.0.0
yt-dlp==2024.3.10

# Audio Processing
soundfile==0.12.1
pydub==0.25.1
noisereduce==2.0.1

# Task Queue
celery==5.3.6
flower==2.0.1

# Storage
boto3==1.34.34
minio==7.2.3

# Monitoring & Logging
prometheus-client==0.19.0
elasticsearch==8.12.1
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.44b0

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0
aiohttp==3.9.3

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0 